requirements.txt
kafka-python==2.3.0
requests==2.31.0
python-dotenv==1.0.0
pyspark==3.5.0
pandas==2.0.0
pyarrow==14.0.0
jupyter==1.0.0